Fixed over 500 warnings on windows builds

+++ source/osal/OSAL_Attribute.cpp
@@ -87,7 +87,7 @@ std::string OSAL::Attribute::getName() {
 }
 
 int OSAL::Attribute::getNumValues() {
-    return value.size();
+    return (int) value.size();
 }
 
 char OSAL::Attribute::getType(int index) {
